added roomtemplate

+++ src/resources.rs
@@ -21,8 +21,8 @@ pub struct Output {
 
 #[derive(Default)]
 pub struct Size {
-	pub width: i32,
-	pub height: i32
+	pub width: i64,
+	pub height: i64
 }
 
 #[derive(Default)]
